Add wait for reindex to FhirClient
